Change the params of network rpc apis from usize to u64 (#1641)

+++ chain/api/src/message.rs
@@ -47,7 +47,7 @@ pub enum ChainRequest {
     GetBlockIds {
         start_number: BlockNumber,
         reverse: bool,
-        max_size: usize,
+        max_size: u64,
     },
     GetBlocks(Vec<HashValue>),
     TPS(Option<BlockNumber>),
